+++ history.md
@@ -1,3 +1,9 @@
+## 5.0.0-pre / 2016-07-05
+
+- new fallbacks api #256
+- rule.applyTo doesn't support fallbacks any more
+- remove compatibility fix for babel 6
+
 ## 4.0.3 / 2016-07-05
 
 - fixed bug when plugins called twice on createStyleSheet #258